refactor: move image build pipeline to orchestrator
Moves ensure_image, build_features_layer, parse_build_options, and
feature resolution from cella-cli into cella-orchestrator::image.
The functions now take &ProgressSender instead of &Progress.

The CLI's image.rs becomes a thin bridge: creates a ProgressSender
channel, calls the orchestrator, and renders events via indicatif.
