+++ crates/cella-orchestrator/Cargo.toml
@@ -5,6 +5,11 @@ edition.workspace = true
 license.workspace = true
 
 [dependencies]
+cella-docker.workspace = true
+cella-features.workspace = true
+hex.workspace = true
+serde_json.workspace = true
+sha2.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
